+++ apps/webstore/src/app/app.module.ts
@@ -30,7 +30,7 @@ import { ProfileComponent } from './pages/auth/profile/profile.component';
 import { ProductComponent } from './pages/product/product.component';
 import { VerificationComponent } from './pages/auth/verification/verification.component';
 import { BasketComponent } from './pages/auth/basket/basket.component';
-import { BasketItemComponent } from './shared/components/items/basket-item/basket-item.component';
+import { FamilyTreeBasketItemComponent } from './shared/components/items/family-tree-basket-item/family-tree-basket-item.component';
 import { PaymentSuccessComponent } from './pages/payment-success/payment-success.component';
 import { NotSignedInComponent } from './pages/product/not-signed-in/not-signed-in.component';
 import { CheckoutComponent } from './pages/auth/checkout/checkout.component';
@@ -68,7 +68,7 @@ import { FamilyTreeMiniatureComponent } from './shared/components/products/famil
     NotSignedInComponent,
     AddToBasketModalComponent,
     BasketComponent,
-    BasketItemComponent,
+    FamilyTreeBasketItemComponent,
     CheckoutComponent,
     FamilyTreeCheckoutItemComponent,
     FamilyTreeDesignComponent,
