+++ apps/webstore/src/app/pages/auth/basket/basket.component.spec.ts
@@ -1,7 +1,7 @@
 import { HttpClientModule } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { BasketItemComponent } from '../../../shared/components/items/basket-item/basket-item.component';
+import { BasketItemComponent } from '../../../shared/components/items/family-tree-basket-item/family-tree-basket-item.component';
 import { BasketComponent } from './basket.component';
 
 describe('BasketComponent', () => {
